feat: dashboard up

+++ apps/web/src/lib/db/validators.ts
@@ -82,7 +82,7 @@ export const insertBattleRecordSchema = z.object({
   result: z.enum(["win", "loss", "draw"]),
   myTeam: z.array(z.record(z.string(), z.unknown())).max(6),
   mySelection: z.array(z.number().int().min(0).max(5)).nullish(),
-  rating: z.number().int().min(0).max(100000).nullish(),
+  rating: z.number().min(0).max(100000).nullish(),
   notes: z.string().nullish(),
   playedAt: z.date().nullish(),
   createdAt: z.date().optional(),
